pyproject.toml
[project]
name = "zerodaemon"
version = "0.1.0"
description = "Local AI-driven DevSecOps assistant"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiosqlite>=0.20",
    "pyyaml>=6.0",
    "langgraph>=0.2",
    "langgraph-checkpoint-sqlite>=2.0",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-ollama>=0.2",
    "langchain-openai>=0.2",
    "langchain-google-genai>=2.0",
    "python-nmap>=0.7",
    "ipwhois>=1.3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "duckduckgo-search>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["zerodaemon"]
